dashboard/pyproject.toml
[project]
name = "trinity-dashboard"
version = "0.1.0"
description = "Trinity Bot Dashboard - Web interface for monitoring Telegram channels and pipelines"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "gunicorn>=21.0.0",
]
